+++ node/src/BaseClient.ts
@@ -235,7 +235,11 @@ import {
     TimeoutError,
 } from "./Errors";
 import { GlideClientConfiguration } from "./GlideClient";
-import { GlideClusterClientConfiguration } from "./GlideClusterClient";
+import {
+    GlideClusterClientConfiguration,
+    RouteOption,
+    Routes,
+} from "./GlideClusterClient";
 import { Logger } from "./Logger";
 import {
     command_request,
@@ -502,10 +506,6 @@ export type ScriptOptions = {
      * The arguments for the script.
      */
     args?: GlideString[];
-    /**
-     * {@link Decoder} type which defines how to handle the responses. If not set, the default decoder from the client config will be used.
-     */
-    decoder?: Decoder;
 };
 
 function getRequestErrorClass(
@@ -530,16 +530,93 @@ function getRequestErrorClass(
     return RequestError;
 }
 
+/**
+ * @internal
+ */
+function toProtobufRoute(
+    route: Routes | undefined,
+): command_request.Routes | undefined {
+    if (!route) {
+        return undefined;
+    }
+
+    if (route === "allPrimaries") {
+        return command_request.Routes.create({
+            simpleRoutes: command_request.SimpleRoutes.AllPrimaries,
+        });
+    } else if (route === "allNodes") {
+        return command_request.Routes.create({
+            simpleRoutes: command_request.SimpleRoutes.AllNodes,
+        });
+    } else if (route === "randomNode") {
+        return command_request.Routes.create({
+            simpleRoutes: command_request.SimpleRoutes.Random,
+        });
+    } else if (route.type === "primarySlotKey") {
+        return command_request.Routes.create({
+            slotKeyRoute: command_request.SlotKeyRoute.create({
+                slotType: command_request.SlotTypes.Primary,
+                slotKey: route.key,
+            }),
+        });
+    } else if (route.type === "replicaSlotKey") {
+        return command_request.Routes.create({
+            slotKeyRoute: command_request.SlotKeyRoute.create({
+                slotType: command_request.SlotTypes.Replica,
+                slotKey: route.key,
+            }),
+        });
+    } else if (route.type === "primarySlotId") {
+        return command_request.Routes.create({
+            slotKeyRoute: command_request.SlotIdRoute.create({
+                slotType: command_request.SlotTypes.Primary,
+                slotId: route.id,
+            }),
+        });
+    } else if (route.type === "replicaSlotId") {
+        return command_request.Routes.create({
+            slotKeyRoute: command_request.SlotIdRoute.create({
+                slotType: command_request.SlotTypes.Replica,
+                slotId: route.id,
+            }),
+        });
+    } else if (route.type === "routeByAddress") {
+        let port = route.port;
+        let host = route.host;
+
+        if (port === undefined) {
+            const split = host.split(":");
+
+            if (split.length !== 2) {
+                throw new RequestError(
+                    "No port provided, expected host to be formatted as `{hostname}:{port}`. Received " +
+                        host,
+                );
+            }
+
+            host = split[0];
+            port = Number(split[1]);
+        }
+
+        return command_request.Routes.create({
+            byAddressRoute: { host, port },
+        });
+    }
+}
+
 export type PubSubMsg = {
     message: string;
     channel: string;
     pattern?: string | null;
 };
 
-export type WritePromiseOptions = {
-    decoder?: Decoder;
-    route?: command_request.Routes;
-};
+/**
+ * @internal
+ * A type to combine RouterOption and DecoderOption to be used for creating write promises for the command.
+ * See - {@link DecoderOption} and {@link RouteOption}
+ */
+export type WritePromiseOptions = RouteOption & DecoderOption;
+
 export class BaseClient {
     private socket: net.Socket;
     private readonly promiseCallbackFunctions: [
@@ -754,8 +831,9 @@ export class BaseClient {
             | command_request.ScriptInvocation,
         options: WritePromiseOptions = {},
     ): Promise<T> {
-        const { decoder = this.defaultDecoder, route } = options;
-        const stringDecoder = decoder === Decoder.String ? true : false;
+        const route = toProtobufRoute(options?.route);
+        const stringDecoder: boolean =
+            (options?.decoder ?? this.defaultDecoder) === Decoder.String;
 
         if (this.isClosed) {
             throw new ClosingError(
@@ -1046,8 +1124,7 @@ export class BaseClient {
      * @see {@link https://valkey.io/commands/get/|valkey.io} for details.
      *
      * @param key - The key to retrieve from the database.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns If `key` exists, returns the value of `key`. Otherwise, return null.
      *
      * @example
@@ -1062,9 +1139,9 @@ export class BaseClient {
      */
     public async get(
         key: GlideString,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString | null> {
-        return this.createWritePromise(createGet(key), { decoder: decoder });
+        return this.createWritePromise(createGet(key), options);
     }
 
     /**
@@ -1093,9 +1170,10 @@ export class BaseClient {
             expiry: "persist" | { type: TimeUnit; duration: number };
         } & DecoderOption,
     ): Promise<GlideString | null> {
-        return this.createWritePromise(createGetEx(key, options?.expiry), {
-            decoder: options?.decoder,
-        });
+        return this.createWritePromise(
+            createGetEx(key, options?.expiry),
+            options,
+        );
     }
 
     /**
@@ -1104,8 +1182,7 @@ export class BaseClient {
      * @see {@link https://valkey.io/commands/getdel/|valkey.io} for details.
      *
      * @param key - The key to retrieve from the database.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns If `key` exists, returns the `value` of `key`. Otherwise, return `null`.
      *
      * @example
@@ -1118,9 +1195,9 @@ export class BaseClient {
      */
     public async getdel(
         key: GlideString,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString | null> {
-        return this.createWritePromise(createGetDel(key), { decoder: decoder });
+        return this.createWritePromise(createGetDel(key), options);
     }
 
     /**
@@ -1135,8 +1212,7 @@ export class BaseClient {
      * @param key - The key of the string.
      * @param start - The starting offset.
      * @param end - The ending offset.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns A substring extracted from the value stored at `key`.
      *
      * @example
@@ -1156,11 +1232,12 @@ export class BaseClient {
         key: GlideString,
         start: number,
         end: number,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString | null> {
-        return this.createWritePromise(createGetRange(key, start, end), {
-            decoder: decoder,
-        });
+        return this.createWritePromise(
+            createGetRange(key, start, end),
+            options,
+        );
     }
 
     /** Set the given key with the given value. Return value is dependent on the passed options.
@@ -1198,9 +1275,7 @@ export class BaseClient {
         value: GlideString,
         options?: SetOptions & DecoderOption,
     ): Promise<"OK" | GlideString | null> {
-        return this.createWritePromise(createSet(key, value, options), {
-            decoder: options?.decoder,
-        });
+        return this.createWritePromise(createSet(key, value, options), options);
     }
 
     /**
@@ -1311,8 +1386,7 @@ export class BaseClient {
      * @remarks When in cluster mode, the command may route to multiple nodes when `keys` map to different hash slots.
      *
      * @param keys - A list of keys to retrieve values for.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns A list of values corresponding to the provided keys. If a key is not found,
      * its corresponding value in the list will be null.
      *
@@ -1327,9 +1401,9 @@ export class BaseClient {
      */
     public async mget(
         keys: GlideString[],
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<(GlideString | null)[]> {
-        return this.createWritePromise(createMGet(keys), { decoder: decoder });
+        return this.createWritePromise(createMGet(keys), options);
     }
 
     /** Set multiple keys to multiple values in a single operation.
@@ -2264,8 +2338,7 @@ export class BaseClient {
      * @see {@link https://valkey.io/commands/lpop/|valkey.io} for details.
      *
      * @param key - The key of the list.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns The value of the first element.
      * If `key` does not exist null will be returned.
      *
@@ -2285,9 +2358,9 @@ export class BaseClient {
      */
     public async lpop(
         key: GlideString,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString | null> {
-        return this.createWritePromise(createLPop(key), { decoder: decoder });
+        return this.createWritePromise(createLPop(key), options);
     }
 
     /** Removes and returns up to `count` elements of the list stored at `key`, depending on the list's length.
@@ -2296,8 +2369,7 @@ export class BaseClient {
      *
      * @param key - The key of the list.
      * @param count - The count of the elements to pop from the list.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns A list of the popped elements will be returned depending on the list's length.
      * If `key` does not exist null will be returned.
      *
@@ -2318,11 +2390,9 @@ export class BaseClient {
     public async lpopCount(
         key: GlideString,
         count: number,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString[] | null> {
-        return this.createWritePromise(createLPop(key, count), {
-            decoder: decoder,
-        });
+        return this.createWritePromise(createLPop(key, count), options);
     }
 
     /** Returns the specified elements of the list stored at `key`.
@@ -2335,8 +2405,7 @@ export class BaseClient {
      * @param key - The key of the list.
      * @param start - The starting point of the range.
      * @param end - The end of the range.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns list of elements in the specified range.
      * If `start` exceeds the end of the list, or if `start` is greater than `end`, an empty list will be returned.
      * If `end` exceeds the actual end of the list, the range will stop at the actual end of the list.
@@ -2367,11 +2436,9 @@ export class BaseClient {
         key: GlideString,
         start: number,
         end: number,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString[]> {
-        return this.createWritePromise(createLRange(key, start, end), {
-            decoder: decoder,
-        });
+        return this.createWritePromise(createLRange(key, start, end), options);
     }
 
     /** Returns the length of the list stored at `key`.
@@ -2405,8 +2472,7 @@ export class BaseClient {
      * @param destination - The key to the destination list.
      * @param whereFrom - The {@link ListDirection} to remove the element from.
      * @param whereTo - The {@link ListDirection} to add the element to.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns The popped element, or `null` if `source` does not exist.
      *
      * @example
@@ -2429,11 +2495,11 @@ export class BaseClient {
         destination: GlideString,
         whereFrom: ListDirection,
         whereTo: ListDirection,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString | null> {
         return this.createWritePromise(
             createLMove(source, destination, whereFrom, whereTo),
-            { decoder: decoder },
+            options,
         );
     }
 
@@ -2453,8 +2519,7 @@ export class BaseClient {
      * @param whereFrom - The {@link ListDirection} to remove the element from.
      * @param whereTo - The {@link ListDirection} to add the element to.
      * @param timeout - The number of seconds to wait for a blocking operation to complete. A value of `0` will block indefinitely.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns The popped element, or `null` if `source` does not exist or if the operation timed-out.
      *
      * @example
@@ -2477,11 +2542,11 @@ export class BaseClient {
         whereFrom: ListDirection,
         whereTo: ListDirection,
         timeout: number,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString | null> {
         return this.createWritePromise(
             createBLMove(source, destination, whereFrom, whereTo, timeout),
-            { decoder: decoder },
+            options,
         );
     }
 
@@ -2632,8 +2697,7 @@ export class BaseClient {
      * @see {@link https://valkey.io/commands/rpop/|valkey.io} for details.
      *
      * @param key - The key of the list.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns The value of the last element.
      * If `key` does not exist null will be returned.
      *
@@ -2653,9 +2717,9 @@ export class BaseClient {
      */
     public async rpop(
         key: GlideString,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString | null> {
-        return this.createWritePromise(createRPop(key), { decoder: decoder });
+        return this.createWritePromise(createRPop(key), options);
     }
 
     /** Removes and returns up to `count` elements from the list stored at `key`, depending on the list's length.
@@ -2664,8 +2728,7 @@ export class BaseClient {
      *
      * @param key - The key of the list.
      * @param count - The count of the elements to pop from the list.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns A list of popped elements will be returned depending on the list's length.
      * If `key` does not exist null will be returned.
      *
@@ -2686,11 +2749,9 @@ export class BaseClient {
     public async rpopCount(
         key: GlideString,
         count: number,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString[] | null> {
-        return this.createWritePromise(createRPop(key, count), {
-            decoder: decoder,
-        });
+        return this.createWritePromise(createRPop(key, count), options);
     }
 
     /** Adds the specified members to the set stored at `key`. Specified members that are already a member of this set are ignored.
@@ -3023,9 +3084,10 @@ export class BaseClient {
         keys: GlideString[],
         options?: DecoderOption,
     ): Promise<Set<GlideString>> {
-        return this.createWritePromise<GlideString[]>(createSUnion(keys), {
-            decoder: options?.decoder,
-        }).then((sunion) => new Set<GlideString>(sunion));
+        return this.createWritePromise<GlideString[]>(
+            createSUnion(keys),
+            options,
+        ).then((sunion) => new Set<GlideString>(sunion));
     }
 
     /**
@@ -3502,7 +3564,7 @@ export class BaseClient {
      * @see {@link https://valkey.io/commands/script-load/|SCRIPT LOAD} and {@link https://valkey.io/commands/evalsha/|EVALSHA} on valkey.io for details.
      *
      * @param script - The Lua script to execute.
-     * @param options - The script option that contains keys and arguments for the script.
+     * @param options - (Optional) See {@link ScriptOptions} and {@link DecoderOption}.
      * @returns A value that depends on the script that was executed.
      *
      * @example
@@ -3518,11 +3580,11 @@ export class BaseClient {
      */
     public async invokeScript(
         script: Script,
-        option?: ScriptOptions,
+        options?: ScriptOptions & DecoderOption,
     ): Promise<ReturnType> {
         const scriptInvocation = command_request.ScriptInvocation.create({
             hash: script.getHash(),
-            keys: option?.keys?.map((item) => {
+            keys: options?.keys?.map((item) => {
                 if (typeof item === "string") {
                     // Convert the string to a Buffer
                     return Buffer.from(item);
@@ -3531,7 +3593,7 @@ export class BaseClient {
                     return item;
                 }
             }),
-            args: option?.args?.map((item) => {
+            args: options?.args?.map((item) => {
                 if (typeof item === "string") {
                     // Convert the string to a Buffer
                     return Buffer.from(item);
@@ -3541,9 +3603,7 @@ export class BaseClient {
                 }
             }),
         });
-        return this.createWritePromise(scriptInvocation, {
-            decoder: option?.decoder,
-        });
+        return this.createWritePromise(scriptInvocation, options);
     }
 
     /**
@@ -5685,8 +5745,7 @@ export class BaseClient {
      *
      * @param key - The `key` of the list.
      * @param index - The `index` of the element in the list to retrieve.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns - The element at `index` in the list stored at `key`.
      * If `index` is out of range or if `key` does not exist, null is returned.
      *
@@ -5707,11 +5766,9 @@ export class BaseClient {
     public async lindex(
         key: GlideString,
         index: number,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<GlideString | null> {
-        return this.createWritePromise(createLIndex(key, index), {
-            decoder: decoder,
-        });
+        return this.createWritePromise(createLIndex(key, index), options);
     }
 
     /**
@@ -5827,8 +5884,7 @@ export class BaseClient {
      *
      * @param keys - The `keys` of the lists to pop from.
      * @param timeout - The `timeout` in seconds.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns - An `array` containing the `key` from which the element was popped and the value of the popped element,
      * formatted as [key, value]. If no element could be popped and the timeout expired, returns `null`.
      *
@@ -5842,11 +5898,9 @@ export class BaseClient {
     public async brpop(
         keys: GlideString[],
         timeout: number,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<[GlideString, GlideString] | null> {
-        return this.createWritePromise(createBRPop(keys, timeout), {
-            decoder: decoder,
-        });
+        return this.createWritePromise(createBRPop(keys, timeout), options);
     }
 
     /** Blocking list pop primitive.
@@ -5860,8 +5914,7 @@ export class BaseClient {
      *
      * @param keys - The `keys` of the lists to pop from.
      * @param timeout - The `timeout` in seconds.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns - An `array` containing the `key` from which the element was popped and the value of the popped element,
      * formatted as [key, value]. If no element could be popped and the timeout expired, returns `null`.
      *
@@ -5874,11 +5927,9 @@ export class BaseClient {
     public async blpop(
         keys: GlideString[],
         timeout: number,
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<[GlideString, GlideString] | null> {
-        return this.createWritePromise(createBLPop(keys, timeout), {
-            decoder: decoder,
-        });
+        return this.createWritePromise(createBLPop(keys, timeout), options);
     }
 
     /** Adds all elements to the HyperLogLog data structure stored at the specified `key`.
@@ -6043,8 +6094,7 @@ export class BaseClient {
      * @param keys - A list of `keys` accessed by the function. To ensure the correct execution of functions,
      *     all names of keys that a function accesses must be explicitly provided as `keys`.
      * @param args - A list of `function` arguments and it should not represent names of keys.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns The invoked function's return value.
      *
      * @example
@@ -6057,11 +6107,9 @@ export class BaseClient {
         func: GlideString,
         keys: GlideString[],
         args: GlideString[],
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<ReturnType> {
-        return this.createWritePromise(createFCall(func, keys, args), {
-            decoder,
-        });
+        return this.createWritePromise(createFCall(func, keys, args), options);
     }
 
     /**
@@ -6075,8 +6123,7 @@ export class BaseClient {
      * @param keys - A list of `keys` accessed by the function. To ensure the correct execution of functions,
      *     all names of keys that a function accesses must be explicitly provided as `keys`.
      * @param args - A list of `function` arguments and it should not represent names of keys.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
-     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @param options - (Optional) See {@link DecoderOption}.
      * @returns The invoked function's return value.
      *
      * @example
@@ -6090,11 +6137,12 @@ export class BaseClient {
         func: GlideString,
         keys: GlideString[],
         args: GlideString[],
-        decoder?: Decoder,
+        options?: DecoderOption,
     ): Promise<ReturnType> {
-        return this.createWritePromise(createFCallReadOnly(func, keys, args), {
-            decoder,
-        });
+        return this.createWritePromise(
+            createFCallReadOnly(func, keys, args),
+            options,
+        );
     }
 
     /**
@@ -6262,7 +6310,7 @@ export class BaseClient {
     ): Promise<[GlideString, [number?, number?, [number, number]?]?][]> {
         return this.createWritePromise(
             createGeoSearch(key, searchFrom, searchBy, options),
-            { decoder: options?.decoder },
+            options,
         );
     }
 
@@ -6957,9 +7005,10 @@ export class BaseClient {
     public async pubsubChannels(
         options?: { pattern?: GlideString } & DecoderOption,
     ): Promise<GlideString[]> {
-        return this.createWritePromise(createPubSubChannels(options?.pattern), {
-            decoder: options?.decoder,
-        });
+        return this.createWritePromise(
+            createPubSubChannels(options?.pattern),
+            options,
+        );
     }
 
     /**